+++ aiohttp/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '2.0.1dev'
+__version__ = '2.1.0dev'
 
 # This relies on each of the submodules having an __all__ variable.