+++ CHANGES.rst
@@ -1,8 +1,11 @@
 Changes
 =======
 
-2.0.1 (2017-xx-xx)
+
+2.1.0 (2017-xx-xx)
 ------------------
+
+- Content disposition with semicolon in filename #917
 
 - Added `request_info` to response object and `ClientResponseError`.